+++ mjolnir/input/read_external_potential.hpp
@@ -44,7 +44,7 @@ read_lennard_jones_wall_potential(const toml::value& external)
     for(const auto& param : ps)
     {
         const auto idx = find_parameter<std::size_t>(param, env, "index") +
-                         find_parameter_or<std::size_t>(param, env, "offset", 0);
+                         find_parameter_or<std::int64_t>(param, env, "offset", 0);
         const auto s   = find_parameter<real_type>(param, env, "sigma", u8"σ");
         const auto e   = find_parameter<real_type>(param, env, "epsilon", u8"ε");
         params.emplace_back(idx, std::make_pair(s, e));
@@ -84,7 +84,7 @@ read_excluded_volume_wall_potential(const toml::value& external)
     for(const auto& param : ps)
     {
         const auto idx = find_parameter<std::size_t>(param, env, "index") +
-                         find_parameter_or<std::size_t>(param, env, "offset", 0);
+                         find_parameter_or<std::int64_t>(param, env, "offset", 0);
         const auto rad = find_parameter<real_type>(param, env, "radius");
         params.emplace_back(idx, rad);
         MJOLNIR_LOG_INFO("idx = ", idx, ", radius = ", rad);
@@ -127,7 +127,7 @@ read_implicit_membrane_potential(const toml::value& external)
     for(const auto& param : ps)
     {
         const auto idx = find_parameter<std::size_t>(param, env, "index") +
-                         find_parameter_or<std::size_t>(param, env, "offset", 0);
+                         find_parameter_or<std::int64_t>(param, env, "offset", 0);
         const auto h   = find_parameter<real_type  >(param, env, "hydrophobicity");
         params.emplace_back(idx, h);
         MJOLNIR_LOG_INFO("idx = ", idx, ", hydrophobicity = ", h);
